+++ semgrep/semgrep/target_manager.py
@@ -4,7 +4,9 @@ import subprocess
 import sys
 import tempfile
 import time
+from collections import defaultdict
 from pathlib import Path
+from typing import Callable
 from typing import cast
 from typing import Collection
 from typing import Dict
@@ -12,29 +14,46 @@ from typing import FrozenSet
 from typing import Iterable
 from typing import Iterator
 from typing import List
+from typing import Mapping
 from typing import Optional
 from typing import Sequence
 from typing import Set
+from typing import Tuple
+
+# usually this would be a try...except ImportError
+# but mypy understands only this
+# see https://github.com/python/mypy/issues/1393
+if sys.version_info[:2] >= (3, 8):
+    # Literal is available in stdlib since Python 3.8
+    from typing import Literal
+else:
+    from typing_extensions import Literal
 
 import attr
+import click
+from attr import Factory
 from wcmatch import glob as wcglob
 
 from semgrep.config_resolver import resolve_targets
+from semgrep.constants import Colors
 from semgrep.error import FilesNotFoundError
+from semgrep.formatter.text import width
 from semgrep.ignores import FileIgnore
 from semgrep.semgrep_types import FileExtension
 from semgrep.semgrep_types import LANGUAGE
 from semgrep.semgrep_types import Language
 from semgrep.semgrep_types import LanguageDefinition
 from semgrep.semgrep_types import Shebang
+from semgrep.types import FilteredTargets
 from semgrep.util import partition_set
 from semgrep.util import sub_check_output
+from semgrep.util import with_color
 from semgrep.verbose_logging import getLogger
 
 logger = getLogger(__name__)
 
 MAX_CHARS_TO_READ_FOR_SHEBANG = 255
-
+PATHS_ALWAYS_SKIPPED = (".git",)
 
 ALL_EXTENSIONS: Collection[FileExtension] = {
     ext
@@ -74,6 +93,174 @@ def converted_pipe_targets(targets: Sequence[str]) -> Iterator[Sequence[str]]:
 
 
 @attr.s(auto_attribs=True)
+class IgnoreLog:
+    """Keeps track of which paths were ignored for what reason.
+
+    Each attribute is a distinct reason why files could be ignored.
+
+    Some reason can apply once per rule; these are mappings keyed on the rule id.
+    """
+
+    target_manager: "TargetManager"
+
+    semgrepignored: Set[Path] = Factory(set)
+    always_skipped: Set[Path] = Factory(set)
+    cli_includes: Set[Path] = Factory(set)
+    cli_excludes: Set[Path] = Factory(set)
+    size_limit: Set[Path] = Factory(set)
+
+    rule_includes: Dict[str, Set[Path]] = Factory(lambda: defaultdict(set))
+    rule_excludes: Dict[str, Set[Path]] = Factory(lambda: defaultdict(set))
+
+    @property
+    def rule_ids_with_skipped_paths(self) -> FrozenSet[str]:
+        """All rule IDs that have skipped paths.
+
+        Note that if a rule ID defines excludes/includes,
+        but they didn't skip any paths,
+        that rule ID will not show up here.
+        """
+        return frozenset(
+            {
+                *(rule_id for rule_id, skips in self.rule_includes.items() if skips),
+                *(rule_id for rule_id, skips in self.rule_excludes.items() if skips),
+            }
+        )
+
+    def __str__(self) -> str:
+        skip_fragments = []
+
+        if self.target_manager.respect_git_ignore:
+            skip_fragments.append("all .gitignored files")
+
+        if self.cli_includes:
+            skip_fragments.append(
+                f"{len(self.cli_includes)} files not matching --include patterns"
+            )
+        if self.cli_excludes:
+            skip_fragments.append(
+                f"{len(self.cli_excludes)} files matching --exclude patterns"
+            )
+        if self.size_limit:
+            skip_fragments.append(
+                f"{len(self.size_limit)} files larger than {self.target_manager.max_target_bytes / 1000 / 1000} MB"
+            )
+        if self.semgrepignored:
+            skip_fragments.append(
+                f"{len(self.semgrepignored)} files matching .semgrepignore patterns"
+            )
+
+        if not skip_fragments:
+            return "no files were skipped"
+
+        message = "skipped: " + ", ".join(skip_fragments)
+
+        message += "\nfor a detailed list of skipped files, run semgrep with the --verbose flag\n"
+        return message
+
+    def yield_verbose_lines(self) -> Iterator[Tuple[Literal[0, 1, 2], str]]:
+        """Yields lines of verbose output for the skipped files.
+
+        The returned tuple is (level, message).
+        The level is a number; one of 0, 1, or 2, which sets the indentation when outputting the line.
+        """
+        yield 0, "Files skipped:"
+
+        yield 1, "Always skipped by Semgrep:"
+        if self.always_skipped:
+            for path in self.always_skipped:
+                yield 2, with_color(Colors.cyan, str(path))
+        else:
+            yield 2, "<none>"
+
+        yield 1, "Skipped by .gitignore:"
+        if self.target_manager.respect_git_ignore:
+            yield 1, "(Disable by passing --no-git-ignore)"
+            yield 2, "<all files not listed by `git ls-files` were skipped>"
+        else:
+            yield 1, "(Disabled with --no-git-ignore)"
+            yield 2, "<none>"
+
+        yield 1, "Skipped by .semgrepignore:"
+        yield 1, "(See: https://semgrep.dev/docs/ignoring-files-folders-code/#understanding-semgrep-defaults)"
+        if self.semgrepignored:
+            for path in self.semgrepignored:
+                yield 2, with_color(Colors.cyan, str(path))
+        else:
+            yield 2, "<none>"
+
+        yield 1, "Skipped by --include patterns:"
+        if self.cli_includes:
+            for path in self.cli_includes:
+                yield 2, with_color(Colors.cyan, str(path))
+        else:
+            yield 2, "<none>"
+
+        yield 1, "Skipped by --exclude patterns:"
+        if self.cli_excludes:
+            for path in self.cli_excludes:
+                yield 2, with_color(Colors.cyan, str(path))
+        else:
+            yield 2, "<none>"
+
+        yield 1, f"Skipped by limiting to files smaller than {self.target_manager.max_target_bytes} bytes:"
+        yield 1, "(Adjust with the --max-target-bytes flag)"
+        if self.size_limit:
+            for path in self.size_limit:
+                yield 2, with_color(Colors.cyan, str(path))
+        else:
+            yield 2, "<none>"
+
+        for rule_id in self.rule_ids_with_skipped_paths:
+            if rule_id.startswith("fingerprints."):
+                # Skip fingerprint rules, since they all have include patterns
+                continue
+            yield 1, f"Skipped only for {with_color(Colors.bright_blue, rule_id)} by the rule's include patterns:"
+            if self.rule_includes[rule_id]:
+                for path in self.rule_includes[rule_id]:
+                    yield 2, with_color(Colors.cyan, str(path))
+            else:
+                yield 2, "<none>"
+
+            yield 1, f"Skipped only for {with_color(Colors.bright_blue, rule_id)} by the rule's exclude patterns:"
+            if self.rule_excludes[rule_id]:
+                for path in self.rule_excludes[rule_id]:
+                    yield 2, with_color(Colors.cyan, str(path))
+            else:
+                yield 2, "<none>"
+
+    def verbose_output(self) -> str:
+        formatters_by_level: Mapping[int, Callable[[str], str]] = {
+            0: lambda line: "\n".join([40 * "=", line, 40 * "="]),
+            1: lambda line: click.wrap_text(
+                with_color(Colors.foreground, line, bold=True),
+                width,
+                2 * " ",
+                2 * " ",
+                False,
+            ),
+            2: lambda line: click.wrap_text(
+                line,
+                width,
+                "   • ",
+                "     ",
+                False,
+            ),
+        }
+        output = ""
+
+        prev_level = None
+        for level, line in self.yield_verbose_lines():
+            if prev_level != level:
+                output += "\n"
+            formatter = formatters_by_level[level]
+            output += formatter(line) + "\n"
+            prev_level = level
+
+        return output
+
+
+@attr.s(auto_attribs=True)
 class TargetManager:
     """
     Handles all file include/exclude logic for semgrep
@@ -84,6 +271,8 @@ class TargetManager:
     If skip_unknown_extensions is False then targets with extensions that are
     not understood by semgrep will always be returned by get_files. Else will discard
     targets with unknown extensions
+
+    TargetManager not to be confused with https://jobs.target.com/search-jobs/store%20manager
     """
 
     includes: Sequence[str]
@@ -93,8 +282,9 @@ class TargetManager:
     respect_git_ignore: bool
     skip_unknown_extensions: bool
     file_ignore: Optional[FileIgnore]
+    ignore_log: IgnoreLog = Factory(IgnoreLog, takes_self=True)
 
-    _filtered_targets: Dict[Language, FrozenSet[Path]] = attr.ib(factory=dict)
+    _filtered_targets: Dict[Language, FilteredTargets] = attr.ib(factory=dict)
 
     @targets.validator
     def _check_exists(self, attribute: str, value: Sequence[str]) -> None:
@@ -105,8 +295,8 @@ class TargetManager:
         i.e. does not exist, is a mount/socket etc.
         """
         targets = self.resolve_targets(self.targets)
-        files, _directories = partition_set(lambda p: not p.is_dir(), targets)
-        _explicit_files, nonexistent_files = partition_set(lambda p: p.is_file(), files)
+        files, _ = partition_set(lambda p: not p.is_dir(), targets)
+        _, nonexistent_files = partition_set(lambda p: p.is_file(), files)
         if nonexistent_files:
             raise FilesNotFoundError(tuple(nonexistent_files))
 
@@ -308,52 +498,57 @@ class TargetManager:
 
     @staticmethod
     def filter_includes(
-        arr: FrozenSet[Path], includes: Sequence[str]
-    ) -> FrozenSet[Path]:
+        includes: Sequence[str], *, candidates: FrozenSet[Path]
+    ) -> FilteredTargets:
         """
-        Returns all elements in arr that match any includes pattern
+        Returns all elements in candidates that match any includes pattern
 
-        If includes is empty, returns arr unchanged
+        If includes is empty, returns candidates unchanged
         """
         if not includes:
-            return arr
+            return FilteredTargets(candidates)
 
         includes = TargetManager.preprocess_path_patterns(includes)
         # Need cast b/c types-wcmatch doesn't use generics properly :(
-        return frozenset(
+        kept = frozenset(
             cast(
                 Iterable[Path],
                 wcglob.globfilter(
-                    arr, includes, flags=wcglob.GLOBSTAR | wcglob.DOTGLOB
+                    candidates, includes, flags=wcglob.GLOBSTAR | wcglob.DOTGLOB
                 ),
             )
         )
+        return FilteredTargets(kept, frozenset(candidates - kept))
 
     @staticmethod
     def filter_excludes(
-        arr: FrozenSet[Path], excludes: Sequence[str]
-    ) -> FrozenSet[Path]:
+        excludes: Sequence[str], *, candidates: FrozenSet[Path]
+    ) -> FilteredTargets:
         """
-        Returns all elements in arr that do not match any excludes pattern
+        Returns all elements in candidates that do not match any excludes pattern
 
-        If excludes is empty, returns arr unchanged
+        If excludes is empty, returns candidates unchanged
         """
         if not excludes:
-            return arr
+            return FilteredTargets(candidates)
 
         excludes = TargetManager.preprocess_path_patterns(excludes)
+
         # Need cast b/c types-wcmatch doesn't use generics properly :(
-        return arr - frozenset(
+        removed = frozenset(
             cast(
                 Iterable[Path],
                 wcglob.globfilter(
-                    arr, excludes, flags=wcglob.GLOBSTAR | wcglob.DOTGLOB
+                    candidates, excludes, flags=wcglob.GLOBSTAR | wcglob.DOTGLOB
                 ),
             )
         )
+        return FilteredTargets(frozenset(candidates - removed), removed)
 
     @staticmethod
-    def filter_by_size(arr: FrozenSet[Path], max_target_bytes: int) -> FrozenSet[Path]:
+    def filter_by_size(
+        max_target_bytes: int, *, candidates: FrozenSet[Path]
+    ) -> FilteredTargets:
         """
         Return all the files whose size doesn't exceed the limit.
 
@@ -362,16 +557,17 @@ class TargetManager:
         result.
         """
         if max_target_bytes <= 0:
-            return arr
-        else:
-            return frozenset(
-                path
-                for path in arr
-                if TargetManager._is_valid_file(path)
-                and os.path.getsize(path) <= max_target_bytes
-            )
+            return FilteredTargets(candidates)
 
-    def filtered_files(self, lang: Language) -> FrozenSet[Path]:
+        kept, removed = partition_set(
+            lambda path: TargetManager._is_valid_file(path)
+            and os.path.getsize(path) <= max_target_bytes,
+            candidates,
+        )
+
+        return FilteredTargets(kept, removed)
+
+    def filtered_files(self, lang: Language) -> FilteredTargets:
         """
         Return all files that are decendants of any directory in TARGET that have
         an extension matching LANG that match any pattern in INCLUDES and do not
@@ -386,14 +582,37 @@ class TargetManager:
 
         # Non dir/non file should not exist cause of init time validation
         # See _check_exists()
-        targets = self.resolve_targets(self.targets)
-        files, directories = partition_set(lambda p: not p.is_dir(), targets)
+        discovered_paths = self.resolve_targets(self.targets)
+        files, directories = partition_set(lambda p: not p.is_dir(), discovered_paths)
         explicit_files, _ = partition_set(lambda p: p.is_file(), files)
 
-        targets = self.expand_targets(directories, lang, self.respect_git_ignore)
-        targets = self.filter_includes(targets, self.includes)
-        targets = self.filter_excludes(targets, [*self.excludes, ".git"])
-        targets = self.filter_by_size(targets, self.max_target_bytes)
+        discovered_paths = self.expand_targets(
+            directories, lang, self.respect_git_ignore
+        )
+
+        filtered_targets = self.filter_includes(
+            self.includes,
+            candidates=discovered_paths,
+        )
+        self.ignore_log.cli_includes.update(filtered_targets.removed)
+
+        filtered_targets = self.filter_excludes(
+            self.excludes,
+            candidates=filtered_targets.kept,
+        )
+        self.ignore_log.cli_excludes.update(filtered_targets.removed)
+
+        filtered_targets = self.filter_excludes(
+            PATHS_ALWAYS_SKIPPED,
+            candidates=filtered_targets.kept,
+        )
+        self.ignore_log.always_skipped.update(filtered_targets.removed)
+
+        filtered_targets = self.filter_by_size(
+            self.max_target_bytes,
+            candidates=filtered_targets.kept,
+        )
+        self.ignore_log.size_limit.update(filtered_targets.removed)
 
         # Remove explicit_files with known extensions.
         explicit_files_with_lang_extension = frozenset(
@@ -401,7 +620,7 @@ class TargetManager:
             for f in explicit_files
             if (any(f.match(f"*{ext}") for ext in LANGUAGE.definition_by_id[lang].exts))
         )
-        targets = targets.union(explicit_files_with_lang_extension)
+        targets = filtered_targets.kept.union(explicit_files_with_lang_extension)
 
         if not self.skip_unknown_extensions:
             explicit_files_with_unknown_extensions = frozenset(
@@ -411,14 +630,25 @@ class TargetManager:
             )
             targets = targets.union(explicit_files_with_unknown_extensions)
 
-        targets = (
-            self.file_ignore.filter_paths(targets) if self.file_ignore else targets
+        filtered_targets = (
+            self.file_ignore.filter_paths(candidates=targets)
+            if self.file_ignore
+            else FilteredTargets(targets)
         )
-        self._filtered_targets[lang] = targets
+        self.ignore_log.semgrepignored.update(filtered_targets.removed)
+
+        self._filtered_targets[lang] = FilteredTargets(
+            filtered_targets.kept,
+            discovered_paths - filtered_targets.kept,
+        )
         return self._filtered_targets[lang]
 
     def get_files(
-        self, lang: Language, includes: Sequence[str], excludes: Sequence[str]
+        self,
+        lang: Language,
+        includes: Sequence[str],
+        excludes: Sequence[str],
+        rule_id: str,
     ) -> FrozenSet[Path]:
         """
         Returns list of files that should be analyzed for a LANG
@@ -431,8 +661,21 @@ class TargetManager:
         in TARGET will bypass this global INCLUDE/EXCLUDE filter. The local INCLUDE/EXCLUDE
         filter is then applied.
         """
-        targets = self.filtered_files(lang)
-        targets = self.filter_includes(targets, includes)
-        targets = self.filter_excludes(targets, excludes)
-        targets = self.filter_by_size(targets, self.max_target_bytes)
-        return targets
+        filtered_targets = self.filtered_files(lang)
+
+        filtered_targets = self.filter_includes(
+            includes, candidates=filtered_targets.kept
+        )
+        self.ignore_log.rule_includes[rule_id].update(filtered_targets.removed)
+
+        filtered_targets = self.filter_excludes(
+            excludes, candidates=filtered_targets.kept
+        )
+        self.ignore_log.rule_excludes[rule_id].update(filtered_targets.removed)
+
+        filtered_targets = self.filter_by_size(
+            self.max_target_bytes, candidates=filtered_targets.kept
+        )
+        self.ignore_log.size_limit.update(filtered_targets.removed)
+
+        return filtered_targets.kept